removed some unnecessary casts

+++ Compiler.h
@@ -295,11 +295,11 @@ protected:
 	#define ChangeInt16InCodePage (this->*ChangeInt16InCodePage)
 	inline void ChangeInt16InCodePageB(const int16_t value, const uint32_t index)
 	{
-		*(int16_t*)(CodePageData.data() + index) = Utils::Bitwise::SwapEndian((int16_t)value);
+		*(int16_t*)(CodePageData.data() + index) = Utils::Bitwise::SwapEndian(value);
 	}
 	inline void ChangeInt16InCodePageL(const int16_t value, const uint32_t index)
 	{
-		*(int16_t*)(CodePageData.data() + index) = (int16_t)value;
+		*(int16_t*)(CodePageData.data() + index) = value;
 	}
 	void (CompileBase::*ChangeInt24InCodePage)(const uint32_t value, const uint32_t index);
 	#define ChangeInt24InCodePage (this->*ChangeInt24InCodePage)
